requirements.txt
arrow==0.10.0
binaryornot==0.4.3
chardet==3.0.2
click==6.7
cookiecutter==1.5.1
Django==1.11
django-crispy-forms==1.6.1
django-mptt==0.8.7
django-mptt-admin==0.4.6
future==0.16.0
Jinja2==2.9.6
jinja2-time==0.2.0
MarkupSafe==1.0
money==1.3.0
oauthlib==2.0.2
olefile==0.44
openexchangerates==0.1.1
Pillow==4.1.0
poyo==0.4.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
simplejson==3.10.0
six==1.10.0
whichcraft==0.4.1
